Sidebar first draft
The sidebar is the first stateful component. It handles the state of the
currently selected section

A Layout higher order component is included. It's first component is the
SideBar which is stateless. It manages which NavigationItem is
currently selected. It passes the currently selected item name and a
handler to the NavigationItem via props.