budget header draft
the header for the budget page contains a date picker and the
amount remaining to be budgeted

The file extention of the files containing jsx has
been changed to *.jsx.
As for the header it contains a date picker that will allow
navigation between different months, and the amount that remains
to be budgetted.