Account component first draft
Account renders transactions

Account renders transactions, the state of which lies in the Main
component. A method allowing to toggle the clearance of a given
Transaction is passed in through props to the Transaction component.

The Transaction component also implement the `shouldComponentUpdate`
lifecycle methods which prevents reloading all the Transaction
components each time a Transaction is cleared.