introducting docker
adds a Dockerfile and a docker-compose.yml

The Dockerfile allows to create containers running the application.
Thanks to the docker-compose.yml file, we can now start both the front
application and the api from the front application root folder with one
command. The only thing necessary to achieve that is to have the api's
folder at the same level as the front application's folder and the api
folder should be named `mynab_api`. Enjoy!